Add Sample Modulator Feature
Adds a new feature. The user can pick a sample as in the sample player, but the sensor input gets recorded over a time window. This way, the user has control over the speed at which the sample is played back. A hierarchical hidden markov model is used to achieve this behavior.